# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个游戏管理系统的API文档，包含前端用户、后台管理、游戏管理等功能",
        "title": "游戏管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8801",
    "basePath": "/",
    "paths": {
...
        "/v1/admin/publishers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页获取游戏厂商列表，支持按名称筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-厂商管理"
                ],
                "summary": "获取游戏厂商列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "厂商名称（模糊查询）",
                        "name": "publisher_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态（0停用 1正常）",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page_num",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.PageResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "records": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.PublisherRes"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的游戏厂商信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-厂商管理"
                ],
                "summary": "创建游戏厂商",
                "parameters": [
                    {
                        "description": "厂商信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PublisherReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/v1/admin/publishers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取游戏厂商的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-厂商管理"
                ],
                "summary": "获取游戏厂商详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "厂商ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PublisherRes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "404": {
                        "description": "厂商不存在",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID更新游戏厂商信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-厂商管理"
                ],
                "summary": "更新游戏厂商",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "厂商ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "厂商信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PublisherReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "404": {
                        "description": "厂商不存在",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID删除游戏厂商",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-厂商管理"
                ],
                "summary": "删除游戏厂商",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "厂商ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "404": {
                        "description": "厂商不存在",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.PageResponse": {
            "type": "object",
            "properties": {
                "current": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "pages": {
                    "description": "总页数",
                    "type": "integer"
                },
                "records": {
                    "description": "数据列表"
                },
                "size": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
...
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据集"
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "dto.PublisherReq": {
            "type": "object",
            "required": [
                "publisher_name",
                "status"
            ],
            "properties": {
                "description": {
                    "description": "厂商介绍，可选",
                    "type": "string"
                },
                "founded_date": {
                    "description": "成立日期，可选",
                    "type": "string"
                },
                "logo_url": {
                    "description": "厂商LOGO，可选",
                    "type": "string"
                },
                "publisher_id": {
                    "description": "厂商ID，更新时需要，创建时忽略",
                    "type": "integer"
                },
                "publisher_name": {
                    "description": "厂商名称，必填",
                    "type": "string"
                },
                "status": {
                    "description": "状态，必填，只能是0或1",
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "website": {
                    "description": "官方网站，可选",
                    "type": "string"
                }
            }
        },
        "dto.PublisherRes": {
            "type": "object",
            "properties": {
                "create_by": {
                    "description": "创建者",
                    "type": "string"
                },
                "create_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "description": {
                    "description": "厂商介绍",
                    "type": "string"
                },
                "founded_date": {
                    "description": "成立日期",
                    "type": "string"
                },
                "logo_url": {
                    "description": "厂商LOGO",
                    "type": "string"
                },
                "publisher_id": {
                    "description": "厂商ID",
                    "type": "integer"
                },
                "publisher_name": {
                    "description": "厂商名称",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "integer"
                },
                "update_by": {
                    "description": "更新者",
                    "type": "string"
                },
                "update_time": {
                    "description": "更新时间",
                    "type": "string"
                },
                "website": {
                    "description": "官方网站",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}